refactor(extension: podman): move windows related checks to `src/checks/windows`
