+++ extensions/podman/packages/extension/src/checks/windows/hyperv-podman-version-check.ts
@@ -18,8 +18,8 @@
 import type { CheckResult } from '@podman-desktop/api';
 import { compareVersions } from 'compare-versions';
 
-import { getPodmanInstallation } from '../utils/podman-cli';
-import { BaseCheck } from './base-check';
+import { getPodmanInstallation } from '../../utils/podman-cli';
+import { BaseCheck } from '../base-check';
 
 export class HyperVPodmanVersionCheck extends BaseCheck {
   title = 'Minimum Podman Version for Hyper-V';
