+++ extensions/podman/packages/extension/src/checks/windows/virtual-machine-platform-check.spec.ts
