+++ extensions/podman/packages/extension/src/extension.ts
@@ -30,11 +30,11 @@ import { compareVersions } from 'compare-versions';
 import type { PodmanExtensionApi, PodmanRunOptions } from '../../api/src/podman-extension-api';
 import { SequenceCheck } from './checks/base-check';
 import { getDetectionChecks } from './checks/detection-checks';
-import { HyperVCheck } from './checks/hyperv-check';
-import { HyperVPodmanVersionCheck } from './checks/hyperv-podman-version-check';
 import { MacKrunkitPodmanMachineCreationCheck, MacPodmanInstallCheck } from './checks/macos-checks';
-import { WSLVersionCheck } from './checks/wsl-version-check';
-import { WSL2Check } from './checks/wsl2-check';
+import { HyperVCheck } from './checks/windows/hyperv-check';
+import { HyperVPodmanVersionCheck } from './checks/windows/hyperv-podman-version-check';
+import { WSLVersionCheck } from './checks/windows/wsl-version-check';
+import { WSL2Check } from './checks/windows/wsl2-check';
 import { PodmanCleanupMacOS } from './cleanup/podman-cleanup-macos';
 import { PodmanCleanupWindows } from './cleanup/podman-cleanup-windows';
 import { KrunkitHelper } from './helpers/krunkit-helper';
