+++ extensions/podman/packages/extension/src/installer/win-installer.ts
@@ -24,13 +24,13 @@ import type { ExtensionContext, InstallCheck, RunError, TelemetryLogger } from '
 import { process as processAPI, ProgressLocation, window } from '@podman-desktop/api';
 
 import { OrCheck, SequenceCheck } from '../checks/base-check';
-import { HyperVCheck } from '../checks/hyperv-check';
-import { VirtualMachinePlatformCheck } from '../checks/virtual-machine-platform-check';
-import { WinBitCheck } from '../checks/win-bit-check';
-import { WinMemoryCheck } from '../checks/win-memory-check';
-import { WinVersionCheck } from '../checks/win-version-check';
-import { WSLVersionCheck } from '../checks/wsl-version-check';
-import { WSL2Check } from '../checks/wsl2-check';
+import { HyperVCheck } from '../checks/windows/hyperv-check';
+import { VirtualMachinePlatformCheck } from '../checks/windows/virtual-machine-platform-check';
+import { WinBitCheck } from '../checks/windows/win-bit-check';
+import { WinMemoryCheck } from '../checks/windows/win-memory-check';
+import { WinVersionCheck } from '../checks/windows/win-version-check';
+import { WSLVersionCheck } from '../checks/windows/wsl-version-check';
+import { WSL2Check } from '../checks/windows/wsl2-check';
 import podman5Json from '../podman5.json';
 import { getAssetsFolder } from '../utils/util';
 import { BaseInstaller } from './base-installer';
